# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A powerful URL shortening service with analytics, QR codes, and link management.",
        "title": "Patrn.ink URL Shortener API",
        "termsOfService": "https://patrn.ink/terms",
        "contact": {
            "name": "API Support",
            "url": "https://patrn.ink/support",
            "email": "support@patrn.ink"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/analytics/{code}/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exports analytics data for a specific link as CSV or JSON",
                "produces": [
                    "application/json",
                    "text/csv"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Export link analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, default: 30 days ago)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, default: today)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Export format: csv or json (default: csv)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV or JSON file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not link owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/bulk/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes or archives multiple links at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk delete/archive links",
                "parameters": [
                    {
                        "description": "Bulk delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BulkDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BulkDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/bulk/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports multiple links at once from a JSON array",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk import links",
                "parameters": [
                    {
                        "description": "Bulk import request (max 100 links)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BulkImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BulkImportResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/export/links": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exports all user's links as CSV or JSON file",
                "produces": [
                    "application/json",
                    "text/csv"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Export all links",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Export format: csv or json (default: csv)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV or JSON file download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BulkDeleteRequest": {
            "type": "object",
            "required": [
                "codes"
            ],
            "properties": {
                "archive": {
                    "description": "If true, archive instead of delete",
                    "type": "boolean"
                },
                "codes": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.BulkDeleteResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.BulkImportError": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.BulkImportItem": {
            "type": "object",
            "required": [
                "long_url"
            ],
            "properties": {
                "custom_code": {
                    "type": "string"
                },
                "long_url": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.BulkImportRequest": {
            "type": "object",
            "required": [
                "links"
            ],
            "properties": {
                "links": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.BulkImportItem"
                    }
                }
            }
        },
        "models.BulkImportResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CreateLinkResponse"
                    }
                },
                "failed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BulkImportError"
                    }
                }
            }
        },
...
        "models.CreateLinkResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "long_url": {
                    "type": "string"
                },
                "qr_code_url": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "short_code": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "API token for programmatic access",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token or API token (prefix with \"Bearer \")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}